factor out _serialize_pod_spec into separate util function
